Move zsh path to own file
`.zprofile` loads the path on the login shell, but once X starts there's
no GDM (Graphical Desktop Manager) to run login scripts in the X
session. Instead we source the path in `.zshrc` for interactive shells.

This implies non-interactive shells will not have the full `$PATH`,
something to improve later.